Refactor API names and locations
